+++ .editorconfig
@@ -1,7 +1,8 @@
+root=true
 [*]
 indent_style = space
 indent_size = 2
 end_of_line = lf
 trim_trailing_whitespace = true
 insert_final_newline = true
-root=true
+
