+++ .eslintrc.json
@@ -7,10 +7,11 @@
     "plugin:prettier/recommended",
     "prettier"
   ],
+  "ignorePatterns": ["/node_modules", "server.js", "/dist"],
   "overrides": [],
   "parser": "@typescript-eslint/parser",
   "parserOptions": { "ecmaVersion": "latest", "sourceType": "module" },
-  "plugins": ["@typescript-eslint", "prettier"],
+  "plugins": ["@typescript-eslint", "import", "prettier"],
   "rules": {
     "max-len": [2, 100],
     "@typescript-eslint/no-unused-vars": 2,
@@ -25,6 +26,12 @@
       "error",
       { "devDependencies": ["**/*.test.ts", "./*.config.js"] }
     ]
+
   },
-  "settings": { "import/resolver": { "node": { "extensions": [".js", ".ts"] } } }
+  "settings": {
+    "import/parsers": {
+			"@typescript-eslint/parser": [".ts"]
+		},
+     "import/resolver": { "node": { "extensions": [".js", ".ts"] } }
+    }
 }
