+++ src/api/interfaces.ts
@@ -35,3 +35,22 @@ export interface IChat {
     content: string;
   };
 }
+
+interface IFile {
+  id: number;
+  user_id: number;
+  path: string;
+  filename: string;
+  content_type: string;
+  content_size: number;
+  upload_date: string;
+}
+
+export interface IMessage {
+  chat_id: number;
+  time: string;
+  type: string;
+  user_id: number;
+  content: string;
+  file?: IFile;
+}
