+++ src/components/AttachInput/attachInput.ts
@@ -1,4 +1,4 @@
-import { Block } from "utils/Block";
+import { Block } from "../../utils/Block";
 import styles from "./attachInput.pcss";
 
 interface IAttachInputProps {
