+++ src/components/Error/error.ts
@@ -1,4 +1,4 @@
-import { Block } from "utils/Block";
+import { Block } from "../../utils/Block";
 import styles from "./error.pcss";
 
 interface IErrorProps {
