+++ src/components/ListWithScroll/listWithScroll.ts
@@ -1,4 +1,4 @@
-import { Block } from "utils/Block";
+import { Block } from "../../utils/Block";
 import styles from "./listWithScroll.pcss";
 
 export class ListWithScroll extends Block {
