+++ src/components/MessageListItem/messageListItem.ts
@@ -1,6 +1,6 @@
-import { IMessage } from "controllers/MessagesController";
-import { Block } from "utils/Block";
-import { formatMessageTime } from "utils/functions/formatMessageTime";
+import { IMessage } from "../../api/interfaces";
+import { Block } from "../../utils/Block";
+import { formatMessageTime } from "../../utils/functions/formatMessageTime";
 import styles from "./messageListItem.pcss";
 
 export interface IMessageListItemProps {
