+++ src/hocs/withChats.ts
@@ -1,5 +1,4 @@
-import { IChat } from "api/interfaces";
-import { IMessage } from "controllers/MessagesController";
+import { IChat, IMessage } from "../api/interfaces";
 import { withStore } from "./withStore";
 
 export const withChats = withStore(
