+++ src/index.ts
@@ -1,14 +1,14 @@
-import { registerComponent } from "utils/registerComponent";
-import * as components from "components";
-import { LoginPage } from "pages/Login";
-import { RegistrationPage } from "pages/Registration";
-import { ChatsPage } from "pages/Chats";
-import { ProfilePage } from "pages/Profile";
-import { router } from "utils/Router";
-import { Routes } from "utils/types/Routes";
-import { controller } from "controllers/AuthController";
-import { Block } from "utils/Block";
-import "styles/global.pcss";
+import { registerComponent } from "./utils/registerComponent";
+import * as components from "./components";
+import { LoginPage } from "./pages/Login";
+import { RegistrationPage } from "./pages/Registration";
+import { ChatsPage } from "./pages/Chats";
+import { ProfilePage } from "./pages/Profile";
+import { router } from "./utils/Router";
+import { Routes } from "./utils/types/Routes";
+import { controller } from "./controllers/AuthController";
+import { Block } from "./utils/Block";
+import "./styles/global.pcss";
 
 window.addEventListener("DOMContentLoaded", async () => {
   Object.values(components).forEach((component) => registerComponent(component as typeof Block));
